:recycle: (Home) extract usePropositions hook and fix types
